services/base/auth-backend/docker/files/requirements.txt
fastapi==0.95.0
pydantic==1.10.7
uvicorn==0.21.1
pyjwt==2.6.0
SQLAlchemy==2.0
psycopg2-binary==2.9.3
sqlalchemy-json==0.6.0
